+++ apps/frontend/src/lib/@types/auth/user.ts
@@ -24,11 +24,3 @@ export interface User {
   accounts: Partial<Record<AccountProvider, AuthAccount>>
   integrations: Partial<Record<IntegrationType, { connected: boolean }>>
 }
-
-// this is the type for showing the integration in profile page
-export interface Integration {
-  key: IntegrationType;
-  icon: JSX.Element;
-  name: string;
-  description: string;
-}
